# docs/spec/basemetadata.openapi.spec.yaml
openapi: 3.0.3
info:
  title: PinShare Metadata Store API
  description: |-
    This API provides access to a CRDT-based metadata store for files in the PinShare system.
    It allows for creating, retrieving, and updating file metadata.
    The system uses a Last-Write-Wins (LWW) strategy for simple fields and simulates
    other CRDT types like OR-Sets for tags and PN-Counters for votes.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /files/{fileSHA256}/download:
    get:
      summary: Download file from IPFS
      operationId: downloadFile
      tags:
        - Metadata
      description: Downloads the file content from IPFS on-demand. The file is fetched from IPFS if not in cache, then streamed to the client with appropriate headers for download.
      parameters:
        - name: fileSHA256
          in: path
          required: true
          description: The SHA256 hash of the file.
          schema:
            type: string
            format: sha256
      responses:
        '200':
          description: File downloaded successfully.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          description: File is banned and cannot be downloaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to fetch file from IPFS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /files/{fileSHA256}/content:
    get:
      summary: Serve file content for preview
      operationId: serveFile
      tags:
        - Metadata
      description: Serves the file content for inline preview (e.g., PDF viewer in browser). The file is fetched from IPFS if not in cache, then streamed with inline content disposition.
      parameters:
        - name: fileSHA256
          in: path
          required: true
          description: The SHA256 hash of the file.
          schema:
            type: string
            format: sha256
      responses:
        '200':
          description: File content served successfully.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          description: File is banned and cannot be viewed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to fetch file from IPFS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - message
      properties:
        code:
          type: integer
          format: int32
          description: An optional application-specific error code.
        message:
          type: string
          description: A human-readable error message.